requirements.txt
discord.py
facebook_scraper